% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCumulative.R
\name{addCumulative}
\alias{addCumulative}
\alias{addCumulative,ANY-method}
\alias{addCumulative,IncucyteExperiment-method}
\title{Add cumulative values to the assays}
\usage{
addCumulative(
  x,
  assay,
  name,
  allowDecrease = FALSE,
  suffix = "cumulative",
  sep = "_"
)

\S4method{addCumulative}{ANY}(
  x,
  assay,
  name,
  allowDecrease = FALSE,
  suffix = "cumulative",
  sep = "_"
)

\S4method{addCumulative}{IncucyteExperiment}(
  x,
  assay,
  name,
  allowDecrease = FALSE,
  suffix = "cumulative",
  sep = "_"
)
}
\arguments{
\item{x}{An IncucyteExpeiment object}

\item{assay}{The assays for which a cumulative sum is required.
Defaults to all existing assays}

\item{name}{Names to give to the new (cumulative) assays}

\item{allowDecrease}{logical vector. Allows the cumulative sum to decrease.
For counts of dead cells, allowing a decrease may not be appropriate as dead
cells cannot revive. However, for growth assays this may be more relevant}

\item{suffix}{If the \code{name} argument is not provided, this suffix will
be added to the original assay name}

\item{sep}{Text separator use to separate the assay name and the suffix}
}
\value{
An IncucyteExperiment object
}
\description{
Add assays with cumulative values to an IncucyteExperiment
}
\details{
After specifying which assays, this calculates a cumulative sum of the
values in each requested assay. This may be useful as assays which measure
death are inherently transient, whilst those which measure growth are
implicitly cumulative
}
