% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay2long.R
\name{assay2long}
\alias{assay2long}
\alias{assay2long,ANY-method}
\alias{assay2long,IncucyteExperiment-method}
\title{Convert assayData to long form}
\usage{
assay2long(x, assay, addColData = TRUE, ...)

\S4method{assay2long}{ANY}(x, assay, addColData = TRUE, ...)

\S4method{assay2long}{IncucyteExperiment}(x, assay, addColData = TRUE)
}
\arguments{
\item{x}{An IncucyteExperiment object}

\item{assay}{One or more of the assays contained in the IncucyteExperiment
object. Defaults to all assays}

\item{addColData}{logical. Include colData in the output}

\item{...}{not used}
}
\value{
A data.frame
}
\description{
Convert assayData to long form with additional colData
}
\details{
Extract each requested assay from an IncucyteExperiment and return a
data.frame in long form including each time-point and sample name.
By default any metadata in the colData slot will be added to the output
}
