% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_incu_ratio.R
\name{calc_incu_ratios}
\alias{calc_incu_ratios}
\title{calc incucyte metrics ratios}
\usage{
calc_incu_ratios(incu_tbl)
}
\arguments{
\item{incu_tbl}{A tibble generated with \code{\link{read_incu}}}
}
\value{
A tibble of incucyte metrics with ratios columns.
}
\description{
Metrics ratios relate different measures from the same or different channels to one another.
}
\examples{
incu_tbl <- read_incu(file = system.file("extdata", c("sample_data_GO_Confluence_percent.txt",
                                                      "sample_data_PO_Confluence_percent.txt",
                                                      "sample_data_GO_Count_image.txt"),
                                         package = "incucyter"),
annotation = system.file("extdata", "sample_data_annotation.tsv",
                          package = "incucyter"))
ratios_tbl <- calc_incu_ratios(incu_tbl)
}

