% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checkAes}
\alias{.checkAes}
\title{Check supplied aesthetics are valid for the IncucyteExperiment}
\usage{
.checkAes(aes, x)
}
\arguments{
\item{aes}{A ggplot aesthetic}

\item{x}{The IncucyteExperiment object to be plotted}
}
\value{
logical(1)
}
\description{
Check supplied aesthetics are valid for the IncucyteExperiment
}
\keyword{internal}
