% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.checkAssays}
\alias{.checkAssays}
\title{Check assays}
\usage{
.checkAssays(l)
}
\arguments{
\item{v}{list of values after importing with .parseIncucyte}
}
\value{
logical(1)
}
\description{
Check the rownames and colnames for multiple assays
}
\keyword{internal}
