% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{.getWellItem}
\alias{.getWellItem}
\title{Extract well metadata}
\usage{
.getWellItem(x, sep = "/")
}
\arguments{
\item{x}{A single well extracted from the complete PlateMap}
}
\description{
Extract well metadata from an XML PlateMap
}
\details{
This collates & reorganises data from within each well in the metadata
}
\keyword{internal}
