% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{.parseIncucyte}
\alias{.parseIncucyte}
\title{Parsing function for Incucyte output}
\usage{
.parseIncucyte(f)
}
\arguments{
\item{f}{File path for an Incucyte output file}
}
\value{
numeric matrix
}
\description{
Parsing function for Incucyte output
}
\keyword{internal}
