% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{.parsePlateMap}
\alias{.parsePlateMap}
\title{Parsing function for Incucyte PlateMap}
\usage{
.parsePlateMap(f, value_sep = "/", .autoGroup = TRUE)
}
\arguments{
\item{f}{File path for an Incucyte PlateMap file}
}
\value{
numeric matrix
}
\description{
Parsing function for Incucyte PlateMap
}
\keyword{internal}
