% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncucyteExperiment.R
\docType{class}
\name{IncucyteExperiment-class}
\alias{IncucyteExperiment-class}
\alias{.IncucyteExpt}
\title{The IncucyteExperiment Class}
\arguments{
\item{f}{A vector or list of file paths containing the observed Incuvyte
data}

\item{map}{File path to the PlateMap file}
}
\description{
The IncucyteExperiment Class
}
\details{
The IncucyteExperiment class extends the SummarizedExperiment class and is
an S4 class
}
