% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncucyteExperiment.R
\name{IncucyteExperiment}
\alias{IncucyteExperiment}
\alias{IncucyteExperiment-class}
\title{The IncucyteExperiment Class}
\usage{
IncucyteExperiment(f, map, autoGroup = TRUE, valueSep = "/")
}
\arguments{
\item{f}{A vector or list of file paths containing the observed Incuvyte
data}

\item{map}{File path to the PlateMap file}

\item{autoGroup}{logical. If TRUE treatment groups will be automatically
created from any PlateMap annotations. Ignored if not PlateMap is provided}

\item{valueSep}{Text separator when combining values across treatment groups}
}
\description{
The IncucyteExperiment Class
}
\details{
The IncucyteExperiment class extends the S4 SummarizedExperiment class.

As these experiments are time courses, rowData will contain the time
information in POSIXct format, along with the Elapsed time from t0.
A final column will contain the Elapsed time scaled to be on [0,1].
This aids with model convergence during GLMM fitting

If a PlateMap file is supplied, metadata from this will be incorporated into
the colData element, otherwise colData will simply contain well IDs.

Additionally if a PlateMap file is provided, the metadata element will
contain the complete well information, including empty wells.
}
