% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_incu.R
\name{norm_incu}
\alias{norm_incu}
\title{normalise incucyte metrics}
\usage{
norm_incu(incu_tbl, ref_time = 72)
}
\arguments{
\item{incu_tbl}{A tibble generated with \code{\link{read_incu}}}

\item{ref_time}{Reference timepoint.}
}
\value{
A normalised tibble of incucyte metrics.
}
\description{
\code{norm_incu()} normalises an incucyte metric to a given vector of reference wells and to a given timepoint. This step requires a \code{Reference} column in your incucyte annotation table indicating comma separated reference wells.
}
\examples{
incu_tbl <- read_incu(file = system.file("extdata", "sample_data_GO_Confluence_percent.txt", package = "incucyter"),
annotation = system.file("extdata", "sample_data_annotation.tsv", package = "incucyter"))
norm_incu(incu_tbl, 72)
}

