% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric.R
\name{plot_metrics}
\alias{plot_metrics}
\title{plot metrics, summarise by condition}
\usage{
plot_metrics(incu_tbl, analysis_job = unique(incu_tbl$Analysis_Job),
  metric = unique(incu_tbl$Metric), color = "Treatment", linetype = NULL,
  label = NULL, alpha = 0.5, summarise = T)
}
\arguments{
\item{incu_tbl}{A tibble generated with \code{\link{read_incu}}}

\item{analysis_job}{Which runs should be plotted (e.g. "Experiment_I")?}

\item{metric}{Which metrics (e.g. "Phase_Confluence")?}

\item{color}{Which columns from your annotation table should be used for coloring?}

\item{label}{Which columns from your annotation table should be used as a text label?}

\item{summarise}{plot mean values instead of single images}
}
\value{
A tibble of incucyte metrics.
}
\description{
\code{plot_microplate()} plots microplate for a given metric.
}
\examples{
incu_tbl <- read_incu(file = system.file("extdata", "sample_data_GO_Confluence_percent.txt",
package = "incucyter"),
annotation = system.file("extdata", "sample_data_annotation.tsv",
package = "incucyter"))
plot_metrics(incu_tbl, color = c("Treatment"), label = c("Treatment"), summarise = T)
}

