% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microplate.R
\name{plot_microplate}
\alias{plot_microplate}
\title{plot microplate graphs}
\usage{
plot_microplate(incu_tbl, analysis_job = incu_tbl$Analysis_Job[1],
  metric = incu_tbl$Metric[1], color = "Treatment", label = NULL,
  summarise = F)
}
\arguments{
\item{incu_tbl}{A tibble generated with \code{\link{read_incu}}}

\item{analysis_job}{Which run should be plotted (e.g. "Experiment_I")?}

\item{metric}{Which metric (e.g. "Phase_Confluence")?}

\item{color}{Which columns from your annotation table should be used for coloring?}

\item{label}{Which columns from your annotation table should be used as a text label?}

\item{summarise}{plot mean values instead of single images}
}
\value{
A tibble of incucyte metrics.
}
\description{
\code{plot_microplate()} plots microplate for a given metric.
}
\examples{
incu_tbl <- read_incu(file = system.file("extdata", "sample_data_GO_Confluence_percent.txt", package = "incucyter"),
                      annotation = system.file("extdata", "sample_data_annotation.tsv", package = "incucyter"))
plot_microplate(incu_tbl, metric = "Green_Confluence", color = c("Treatment"), label = c("Treatment"), summarise = T)
}

