% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssayMerged.R
\name{plotAssayMerged}
\alias{plotAssayMerged}
\alias{plotAssayMerged,ANY-method}
\alias{plotAssayMerged,IncucyteExperiment-method}
\title{Plot the assayData merged across replicate wells}
\usage{
plotAssayMerged(x, assay, ..., trans = "log10", pointsize = 1/3)

\S4method{plotAssayMerged}{ANY}(x, assay, ..., trans = "log10", pointsize = 1/3)

\S4method{plotAssayMerged}{IncucyteExperiment}(x, assay, ..., trans = "log10", pointsize = 1/3)
}
\arguments{
\item{x}{An IncucyteExperiment object}

\item{assay}{The assay to be visualised}

\item{...}{Passed to \code{aes_string()}. If no colour, linetype or other
grouping category is provided, the plot will simply be a summary of all
points in all wells. Any provided categories must correspond to columns
contained in the colData slot of \code{x}}

\item{trans}{The transformation for the y-axis}

\item{pointsize}{The size of points to be drawn}
}
\value{
A ggplot object
}
\description{
Plot the assayData from an IncucyteExperiment object merged
across replicate wells
}
\details{
Visualise all data from the requested assay merged by the column(s) passed
to the line and point aesthetics


The returned object is a ggplot object and can be further modified using
scales, themes and additional geoms.
}
