% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssayRaw.R
\name{plotAssayRaw}
\alias{plotAssayRaw}
\alias{plotAssayRaw,ANY-method}
\alias{plotAssayRaw,IncucyteExperiment-method}
\title{Plot the raw assayData}
\usage{
plotAssayRaw(x, assay, ..., trans = "log10", facet = facet_grid(row ~ col))

\S4method{plotAssayRaw}{ANY}(x, assay, ..., trans = "log10", facet = facet_grid(row ~ col))

\S4method{plotAssayRaw}{IncucyteExperiment}(x, assay, ..., trans = "log10", facet = facet_grid(row ~ col))
}
\arguments{
\item{x}{An IncucyteExperiment object}

\item{assay}{The assay to be visualised}

\item{...}{Passed to \code{aes_string()} in the initial call to
\code{ggplot}. Any aesthetics can be specified here, with values passed to
the aesthetics as a character, and that match a column name from colData.}

\item{trans}{The transformation for the y-axis}

\item{facet}{Faceting to add to the plot}
}
\value{
A ggplot object
}
\description{
Plot the raw assayData from an IncucyteExperiment object
}
\details{
Visualise all data from the requested assay. By default a line plot will
be produced for every image within each well. The plot will also be broken
into facets using the row and column information.

Lines can be coloured by any additional column contained in the colData slot
of the IncucyteExperiment object by setting the argument
\code{colour = colName}, which will be passed to \code{aes_string()} within
the initial call to \code{ggplot}. Any additional columns must be quoted.

Additional information can also be added to the facet headers by adding
the column you wish. For example, if each column contains a unique CellType,
you could specify \code{facet = facet_grid(row~col+CellType)}. Note that any
additional column must also be contained in the colData slot.

The returned object is a ggplot object and can be further modified using
scales, themes and additional geoms.
}
