% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseBy.R
\name{summariseBy}
\alias{summariseBy}
\alias{summariseBy,ANY-method}
\alias{summariseBy,IncucyteExperiment-method}
\title{Summarise Treatments from an IncucyteExperiment object}
\usage{
summariseBy(x, by = c("colData", "plateMap"))

\S4method{summariseBy}{ANY}(x, by = c("colData", "plateMap"))

\S4method{summariseBy}{IncucyteExperiment}(x, by = c("colData", "plateMap"))
}
\arguments{
\item{x}{An IncucyteExperiment object}

\item{by}{Summarise by columns in the colData or plateMap slots}
}
\value{
A grouped tibble
}
\description{
Provide numbers of replicates from treatment groups
}
\details{
This summarises any columns from the requested slot which are not positional
(i.e. row, column etc) and summarises the number of values found for each
(treatment) group found
}
