% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAssayMerged.R
\name{plotAssayMerged}
\alias{plotAssayMerged}
\alias{plotAssayMerged,ANY-method}
\alias{plotAssayMerged,IncucyteExperiment-method}
\title{Plot the assayData merged across replicate wells}
\usage{
plotAssayMerged(
  x,
  assay,
  point_aes,
  line_aes,
  trans = "log10",
  pointsize = 1/3
)

\S4method{plotAssayMerged}{ANY}(x)

\S4method{plotAssayMerged}{IncucyteExperiment}(
  x,
  assay,
  point_aes,
  line_aes,
  trans = "log10",
  pointsize = 1/3
)
}
\arguments{
\item{x}{An IncucyteExperiment object}

\item{assay}{The assay to be visualised}

\item{point_aes}{Plotting aesthetics for points.
Any aesthetics must refer to columns in the colData slot of x.
Any aesthetics provided here will become the grouping categories for
calculating and plotting the mean+se of each category.
Must be specified.}

\item{line_aes}{Plotting aesthetics for lines.
If not provided will default to the same as point_aes}

\item{trans}{The transformation for the y-axis}

\item{pointsize}{The size of points to be drawn}
}
\value{
A ggplot object
}
\description{
Plot the assayData from an IncucyteExperiment object merged
across replicate wells
}
\details{
Visualise all data from the requested assay merged by the column(s) passed
to the line and point aesthetics


The returned object is a ggplot object and can be further modified using
scales, themes and additional geoms.
}
